# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  codeberg_org_azzet_azzetbe_internal_shared.APIResponse:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
...
  codeberg_org_azzet_azzetbe_internal_shared.PaginationMeta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  codeberg_org_azzet_azzetbe_internal_tax.CalculationResponse:
    properties:
      base_amount:
        type: string
      counterparty_entity_id:
        type: string
      created_at:
        type: string
      direction:
        type: string
      faktur_number:
        type: string
      id:
        type: string
      period:
        type: string
      status:
        type: string
      tax_amount:
        type: string
      tax_rate:
        type: number
      tax_type:
        type: string
      transaction_description:
        type: string
      transaction_id:
        type: string
      transaction_number:
        type: string
      transaction_type:
        type: string
      workspace_id:
        type: string
    type: object
  codeberg_org_azzet_azzetbe_internal_tax.DocumentRefResponse:
    properties:
      created_at:
        type: string
      document_id:
        type: string
      document_type:
        type: string
      file_name:
        type: string
      id:
        type: string
      ref_type:
        type: string
    type: object
  codeberg_org_azzet_azzetbe_internal_tax.LinkDocumentRequest:
    properties:
      document_id:
        type: string
      ref_type:
        example: FAKTUR_PAJAK
        type: string
    type: object
  codeberg_org_azzet_azzetbe_internal_tax.PPNSummaryResponse:
    properties:
      dpp_keluaran:
        type: string
      dpp_masukan:
        type: string
      net_ppn:
        type: string
      period:
        type: string
      ppn_keluaran:
        type: string
      ppn_masukan:
        type: string
      transaction_count:
        type: integer
    type: object
  codeberg_org_azzet_azzetbe_internal_tax.PPhSummaryResponse:
    properties:
      period_from:
        type: string
      period_to:
        type: string
      rows:
        items:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.PPhSummaryRow'
        type: array
    type: object
  codeberg_org_azzet_azzetbe_internal_tax.PPhSummaryRow:
    properties:
      count:
        type: integer
      direction:
        type: string
      tax_type:
        type: string
      total_base:
        type: string
      total_tax:
        type: string
    type: object
  codeberg_org_azzet_azzetbe_internal_tax.ProfileResponse:
    properties:
      created_at:
        type: string
      default_pph23_rate:
        type: number
      default_ppn_rate:
        type: number
      ebupot_ready:
        type: boolean
      efaktur_ready:
        type: boolean
      entity_id:
        type: string
      id:
        type: string
      is_ppn_liable:
        type: boolean
      notes:
        type: string
      npwp:
        type: string
      pkp_number:
        type: string
      pph23_enabled:
        type: boolean
      tax_office_code:
        type: string
      tax_status:
        type: string
      updated_at:
        type: string
      workspace_id:
        type: string
    type: object
  codeberg_org_azzet_azzetbe_internal_tax.ReportJobResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      id:
        type: string
      period_from:
        type: string
      period_to:
        type: string
      report_type:
        type: string
      result: {}
      status:
        type: string
    type: object
  codeberg_org_azzet_azzetbe_internal_tax.RequestReportRequest:
    properties:
      period_from:
        example: 2026-01
        type: string
      period_to:
        example: 2026-03
        type: string
      report_type:
        example: TAX_OVERVIEW
        type: string
    type: object
  codeberg_org_azzet_azzetbe_internal_tax.UpsertProfileRequest:
    properties:
      default_pph23_rate:
        example: 0.02
        type: number
      default_ppn_rate:
        example: 0.11
        type: number
      ebupot_ready:
        example: false
        type: boolean
      efaktur_ready:
        example: false
        type: boolean
      is_ppn_liable:
        example: true
        type: boolean
      notes:
        type: string
      npwp:
        example: 01.234.567.8-901.000
        type: string
      pkp_number:
        example: 123.456-78.901234
        type: string
      pph23_enabled:
        example: false
        type: boolean
      tax_office_code:
        example: "061"
        type: string
      tax_status:
        example: PKP
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Enterprise-grade accounting, tax, and finance platform. User routes
    use Bearer JWT from /auth/login. Workspace-scoped routes require X-Workspace-ID
    header. Admin routes under /admin use admin JWT from /admin/auth/login.
  title: Azzet API
  version: "1.0"
paths:
...
  /tax/calculations:
    get:
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Period YYYY-MM
        in: query
        name: period
        type: string
      - description: Tax type filter
        in: query
        name: tax_type
        type: string
      - description: Status filter
        in: query
        name: status
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.CalculationResponse'
                  type: array
                meta:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.PaginationMeta'
              type: object
      security:
      - BearerAuth: []
      summary: List tax calculations
      tags:
      - Tax
  /tax/calculations/{id}:
    get:
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.CalculationResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get tax calculation detail
      tags:
      - Tax
  /tax/calculations/{id}/documents:
    get:
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.DocumentRefResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List documents linked to a tax calculation
      tags:
      - Tax
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: string
      - description: Document link
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.LinkDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.DocumentRefResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Link a document to a tax calculation
      tags:
      - Tax
  /tax/profile:
    get:
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.ProfileResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get workspace tax profile
      tags:
      - Tax
    put:
      consumes:
      - application/json
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Tax profile
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.UpsertProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.ProfileResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update workspace tax profile
      tags:
      - Tax
  /tax/reports:
    get:
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.ReportJobResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List tax report jobs
      tags:
      - Tax
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Report request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.RequestReportRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.ReportJobResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Request async tax report generation
      tags:
      - Tax
  /tax/reports/{id}:
    get:
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Report job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.ReportJobResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get tax report job status
      tags:
      - Tax
  /tax/summary/pph:
    get:
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Period from YYYY-MM
        in: query
        name: period_from
        required: true
        type: string
      - description: Period to YYYY-MM
        in: query
        name: period_to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.PPhSummaryResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get PPh summary for a period range
      tags:
      - Tax
  /tax/summary/ppn:
    get:
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Period YYYY-MM
        in: query
        name: period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_tax.PPNSummaryResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get PPN summary for a period
      tags:
      - Tax
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"